+++ fltk-sys/src/menu.rs
@@ -2104,6 +2104,9 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Menu_Item_do_callback(self_: *mut Fl_Menu_Item, w: *mut Fl_Widget);
+}
+extern "C" {
     pub fn Fl_Menu_Item_user_data(arg1: *const Fl_Menu_Item) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {